apps/api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
sqlalchemy==2.0.34
psycopg[binary]==3.2.12
alembic==1.14.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.5.2
python-multipart==0.0.9
jinja2==3.1.4
email-validator==2.2.0
minio==7.2.18
rq==1.16.2
redis==5.0.4
httpx==0.27.2
